refactor(step3): 생성자에서 START_LINE 지정
